ai_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
azure-ai-documentintelligence>=1.0.0
azure-core>=1.30.0
rapidfuzz>=3.9.0
supabase>=2.4.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
Pillow>=10.3.0
numpy>=1.26.0
onnxruntime>=1.18.0
pypdfium2>=4.30.0
paddleocr>=2.7.0;python_version>="3.8"
paddlepaddle>=2.5.0;python_version>="3.8"
opencv-python-headless>=4.8.0
